support fsevents on darwin, fsnotify when not
